Semester_1/Rout-Art/cms/requirements.txt
customtkinter>=5.0.0,<6.0.0
beautifulsoup4>=4.11.0,<5.0.0
gitpython>=3.1.0,<4.0.0
pyinstaller>=6.0.0,<7.0.0